+++ t/sperl_t_hash.c
@@ -35,51 +35,51 @@ int main()
   // Hash - first same key
   {
     SPerl_HASH* hash = SPerl_HASH_new(sperl, 101);
-    int64_t value1 = 3;
+    int32_t value1 = 3;
     SPerl_HASH_insert_norehash(sperl, hash, "key1", 4, &value1);
 
-    int64_t value1_ret = *(int64_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
+    int32_t value1_ret = *(int32_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
     OK(value1_ret == 3);
 
-    int64_t value2 = 5;
+    int32_t value2 = 5;
     SPerl_HASH_insert_norehash(sperl, hash, "key", 3, &value2);
     
-    int64_t value2_ret = *(int64_t*)SPerl_HASH_search(sperl, hash, "key", 3);
+    int32_t value2_ret = *(int32_t*)SPerl_HASH_search(sperl, hash, "key", 3);
     OK(value2_ret == 5);
   }
   
   // Hash - insert and search
   {
     SPerl_HASH* hash = SPerl_HASH_new(sperl, 101);
-    int64_t value1 = 3;
+    int32_t value1 = 3;
     SPerl_HASH_insert_norehash(sperl, hash, "key1", 4, &value1);
-    int64_t hash_value1 = SPerl_HASH_FUNC_calc_hash(sperl, "key1", 4);
-    int64_t index1 = hash_value1 % 101;
+    int32_t hash_value1 = SPerl_HASH_FUNC_calc_hash(sperl, "key1", 4);
+    int32_t index1 = hash_value1 % 101;
    
-    OK(*(int64_t*)hash->table[index1]->value == 3);
+    OK(*(int32_t*)hash->table[index1]->value == 3);
     OK(hash->entries_length == 1);
 
-    int64_t value2 = 5;
+    int32_t value2 = 5;
     SPerl_HASH_insert_norehash(sperl, hash, "key2", 4, &value2);
-    int64_t hash_value2 = SPerl_HASH_FUNC_calc_hash(sperl, "key2", 4);
-    int64_t index2 = hash_value2 % 101;
+    int32_t hash_value2 = SPerl_HASH_FUNC_calc_hash(sperl, "key2", 4);
+    int32_t index2 = hash_value2 % 101;
     
-    OK(*(int64_t*)hash->table[index2]->value == 5);
+    OK(*(int32_t*)hash->table[index2]->value == 5);
     OK(hash->entries_length == 2);
     
     // Replace
-    int64_t value3 = 7;
+    int32_t value3 = 7;
     SPerl_HASH_insert_norehash(sperl, hash, "key1", 4, &value3);
-    int64_t hash_value3 = SPerl_HASH_FUNC_calc_hash(sperl, "key1", 4);
-    int64_t index3 = hash_value3 % 101;
+    int32_t hash_value3 = SPerl_HASH_FUNC_calc_hash(sperl, "key1", 4);
+    int32_t index3 = hash_value3 % 101;
     
-    OK(*(int64_t*)hash->table[index3]->value == 7);
+    OK(*(int32_t*)hash->table[index3]->value == 7);
     OK(hash->entries_length == 2);
     
     // Search
-    int64_t value1_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
+    int32_t value1_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
     OK(value1_1 == 7);
-    int64_t value2_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
+    int32_t value2_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
     OK(value2_1 == 5);
     void* value3_1 = SPerl_HASH_search(sperl, hash, "none", 4);
     OK(!value3_1);
@@ -88,34 +88,34 @@ int main()
   // Hash - insert link
   {
     SPerl_HASH* hash = SPerl_HASH_new(sperl, 1);
-    int64_t value1 = 3;
+    int32_t value1 = 3;
     SPerl_HASH_insert_norehash(sperl, hash, "key1", 4, &value1);
     
-    OK(*(int64_t*)hash->table[0]->value == 3);
+    OK(*(int32_t*)hash->table[0]->value == 3);
     OK(hash->table[0] == &hash->entries[0]);
     OK(hash->table_capacity == 1);
     OK(hash->entries_length == 1);
     
-    int64_t value2 = 5;
+    int32_t value2 = 5;
     SPerl_HASH_insert_norehash(sperl, hash, "key2", 4, &value2);
     
-    OK(*(int64_t*)hash->entries[hash->table[0]->next_index].value == 5);
+    OK(*(int32_t*)hash->entries[hash->table[0]->next_index].value == 5);
     OK(hash->table_capacity == 1);
     OK(hash->entries_length == 2);
     
-    int64_t value3 = 7;
+    int32_t value3 = 7;
     SPerl_HASH_insert_norehash(sperl, hash, "key3", 4, &value3);
     
-    OK(*(int64_t*)hash->entries[hash->entries[hash->table[0]->next_index].next_index].value == 7);
+    OK(*(int32_t*)hash->entries[hash->entries[hash->table[0]->next_index].next_index].value == 7);
     OK(hash->table_capacity == 1);
     OK(hash->entries_length == 3);
 
     // Search
-    int64_t value1_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
+    int32_t value1_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
     OK(value1_1 == 3);
-    int64_t value2_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
+    int32_t value2_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
     OK(value2_1 == 5);
-    int64_t value3_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key3", 4);
+    int32_t value3_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key3", 4);
     OK(value3_1 == 7);
     
     // free
@@ -126,63 +126,63 @@ int main()
   {
     SPerl_HASH* hash = SPerl_HASH_new(sperl, 5);
     
-    int64_t value1 = 1;
+    int32_t value1 = 1;
     SPerl_HASH_insert(sperl, hash, "key1", 4, &value1);
-    int64_t value2 = 2;
+    int32_t value2 = 2;
     SPerl_HASH_insert(sperl, hash, "key2", 4, &value2);
-    int64_t value3 = 3;
+    int32_t value3 = 3;
     SPerl_HASH_insert(sperl, hash, "key3", 4, &value3);
-    int64_t value4 = 4;
+    int32_t value4 = 4;
     SPerl_HASH_insert(sperl, hash, "key4", 4, &value4);
     
     OK(hash->table_capacity == 5);
     OK(hash->entries_length == 4);
     
     // Rehash
-    int64_t value5 = 5;
+    int32_t value5 = 5;
     SPerl_HASH_insert(sperl, hash, "key5", 4, &value5);
 
     OK(hash->table_capacity == 11);
     OK(hash->entries_length == 5);
 
-    int64_t search_value1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
+    int32_t search_value1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
     OK(search_value1 == 1);
-    int64_t search_value2 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
+    int32_t search_value2 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
     OK(search_value2 == 2);
-    int64_t search_value3 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key3", 4);
+    int32_t search_value3 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key3", 4);
     OK(search_value3 == 3);
-    int64_t search_value4 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key4", 4);
+    int32_t search_value4 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key4", 4);
     OK(search_value4 == 4);
-    int64_t search_value5 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key5", 4);
+    int32_t search_value5 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key5", 4);
     OK(search_value5 == 5);
 
-    int64_t value6 = 6;
+    int32_t value6 = 6;
     SPerl_HASH_insert(sperl, hash, "key6", 4, &value6);
-    int64_t value7 = 7;
+    int32_t value7 = 7;
     SPerl_HASH_insert(sperl, hash, "key7", 4, &value7);
-    int64_t value8 = 8;
+    int32_t value8 = 8;
     SPerl_HASH_insert(sperl, hash, "key8", 4, &value8);
-    int64_t value9 = 9;
+    int32_t value9 = 9;
     SPerl_HASH_insert(sperl, hash, "key9", 4, &value9);
 
     OK(hash->table_capacity == 11);
     OK(hash->entries_length == 9);
     
     // Rehash
-    int64_t value10 = 10;
+    int32_t value10 = 10;
     SPerl_HASH_insert(sperl, hash, "key10", 5, &value10);
     OK(hash->table_capacity == 23);
     OK(hash->entries_length == 10);
     
-    int64_t search_value6 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key6", 4);
+    int32_t search_value6 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key6", 4);
     OK(search_value6 == 6);
-    int64_t search_value7 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key7", 4);
+    int32_t search_value7 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key7", 4);
     OK(search_value7 == 7);
-    int64_t search_value8 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key8", 4);
+    int32_t search_value8 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key8", 4);
     OK(search_value8 == 8);
-    int64_t search_value9 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key9", 4);
+    int32_t search_value9 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key9", 4);
     OK(search_value9 == 9);
-    int64_t search_value10 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key10", 5);
+    int32_t search_value10 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key10", 5);
     OK(search_value10 == 10);
   }
   return 0;
